requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
plyer==2.1.0
cryptography==42.0.2
schedule==1.2.1
requests==2.31.0
Pillow==10.2.0
undetected-chromedriver==3.5.5
customtkinter==5.2.2
redis==5.0.1
prometheus_client==0.19.0
psycopg2-binary==2.9.9
